fixup! JAMES-2544 testing cassandra configuration view changes
